CreatureData추가 StatusEffect임시생성
