--- Assets/Art/Animators/Ex_Controller.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9d7184dc74cca8c458819d7d95c0230b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
