+++ Assets/Scripts/Creature/CreatureData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58c63c789d55d53438a2f1dfcc16eb2e
+guid: fece30e92729fd743a39213f77a544df
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
