+++ Assets/Scripts/Creature/ObjectData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f624a70622561a4fa13981f57996066
+guid: c7ba0db61764a414fb1403a580e287e2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
