+++ Assets/Scripts/StatusEffect.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cb0d20b9b7570a44c9625f643b8676b3
+guid: 7c49f83ab97f04648a3fb4f44ee5e65f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
