+++ Assets/Scripts/StatusEffect/StatusEffect.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ebef4c03306eda34b98aa1801465bb90
+guid: 5145ff569394cb2468eef7b2ba2ceab6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
