--- Assets/Scripts/StatusEffectScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 431d0607669700b47aceb60ed58321be
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
